[SPARK] use single kafka container
