+++ .circleci/continue_config.yml
@@ -485,7 +485,7 @@ jobs:
       - run: |
           sudo update-alternatives --set java /usr/lib/jvm/java-8-openjdk-amd64/jre/bin/java
           sudo update-alternatives --set javac /usr/lib/jvm/java-8-openjdk-amd64/bin/javac
-      - run: ./gradlew --no-daemon integrationTest -x test -Pspark.version=<< parameters.spark-version >> -PintegrationTest.tags=<< parameters.integration-test-tags >>
+      - run: ./gradlew --no-daemon :app:integrationTest -x test -Pspark.version=<< parameters.spark-version >> -PintegrationTest.tags=<< parameters.integration-test-tags >>
       - run: ./gradlew --no-daemon jacocoTestReport
       - store_test_results:
           path: app/build/test-results/integrationTest
