+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/lifecycle/SparkReadWriteIntegTest.java
@@ -44,7 +44,6 @@ import io.openlineage.spark.agent.util.SparkVersionUtils;
 import io.openlineage.spark.agent.util.TestOpenLineageEventHandlerFactory;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -54,9 +53,6 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.Pair;
@@ -65,9 +61,6 @@ import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.FileInputFormat;
 import org.apache.hadoop.mapred.JobConf;
 import org.apache.hadoop.mapred.TextInputFormat;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.common.serialization.StringSerializer;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.AnalysisException;
@@ -78,7 +71,6 @@ import org.apache.spark.sql.SaveMode;
 import org.apache.spark.sql.SparkSession;
 import org.apache.spark.sql.catalyst.expressions.GenericRow;
 import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema;
-import org.apache.spark.sql.types.BinaryType$;
 import org.apache.spark.sql.types.IntegerType$;
 import org.apache.spark.sql.types.LongType$;
 import org.apache.spark.sql.types.Metadata;
@@ -87,7 +79,6 @@ import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.assertj.core.api.ObjectAssert;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
@@ -97,8 +88,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.utility.DockerImageName;
 import scala.Tuple2;
 import scala.collection.immutable.HashMap;
 
@@ -116,9 +105,6 @@ class SparkReadWriteIntegTest {
   private static final String SPARK_3 = "(3.*)";
   private static final String SPARK_VERSION = "spark.version";
 
-  private final KafkaContainer kafkaContainer =
-      new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.0.0"));
-
   @BeforeEach
   public void setUp() {
     reset(MockBigQueryRelationProvider.BIG_QUERY);
@@ -128,13 +114,6 @@ class SparkReadWriteIntegTest {
         .thenReturn("ParentJob");
     when(SparkAgentTestExtension.OPEN_LINEAGE_SPARK_CONTEXT.getJobNamespace())
         .thenReturn("Namespace");
-  }
-
-  @AfterEach
-  public void tearDown() {
-    if (kafkaContainer.isCreated()) {
-      kafkaContainer.stop();
-    }
   }
 
   @Test
@@ -497,103 +476,6 @@ class SparkReadWriteIntegTest {
   }
 
   @Test
-  @Tag("kafka")
-  void testWriteWithKafkaSourceProvider(SparkSession spark)
-      throws InterruptedException, TimeoutException {
-    kafkaContainer.start();
-    StructType schema =
-        new StructType(
-            new StructField[] {
-              new StructField("key", StringType$.MODULE$, false, new Metadata(new HashMap<>())),
-              new StructField("value", BinaryType$.MODULE$, false, new Metadata(new HashMap<>()))
-            });
-
-    spark
-        .createDataFrame(
-            Arrays.asList(
-                new GenericRow(new Object[] {"seven", "seven".getBytes(StandardCharsets.UTF_8)}),
-                new GenericRow(new Object[] {"one", "one".getBytes(StandardCharsets.UTF_8)}),
-                new GenericRow(
-                    new Object[] {"fourteen", "fourteen".getBytes(StandardCharsets.UTF_8)}),
-                new GenericRow(
-                    new Object[] {"sixteen", "sixteen".getBytes(StandardCharsets.UTF_8)})),
-            schema)
-        .write()
-        .format("kafka")
-        .option("topic", "topicA")
-        .option("kafka.bootstrap.servers", kafkaContainer.getBootstrapServers())
-        .save();
-
-    StaticExecutionContextFactory.waitForExecutionEnd();
-    ArgumentCaptor<OpenLineage.RunEvent> lineageEvent =
-        ArgumentCaptor.forClass(OpenLineage.RunEvent.class);
-    Mockito.verify(SparkAgentTestExtension.OPEN_LINEAGE_SPARK_CONTEXT, times(4))
-        .emit(lineageEvent.capture());
-    OpenLineage.RunEvent completeEvent = lineageEvent.getAllValues().get(2);
-    assertThat(completeEvent).hasFieldOrPropertyWithValue(EVENT_TYPE, RunEvent.EventType.COMPLETE);
-    String kafkaNamespace =
-        "kafka://"
-            + kafkaContainer.getHost()
-            + ":"
-            + kafkaContainer.getMappedPort(KafkaContainer.KAFKA_PORT);
-    assertThat(completeEvent.getOutputs())
-        .hasSize(1)
-        .first()
-        .hasFieldOrPropertyWithValue(NAME, "topicA")
-        .hasFieldOrPropertyWithValue(NAMESPACE, kafkaNamespace);
-  }
-
-  @Test
-  @Tag("kafka")
-  void testReadWithKafkaSourceProviderUsingAssignConfig(SparkSession spark)
-      throws InterruptedException, TimeoutException, ExecutionException {
-    kafkaContainer.start();
-    Properties p = new Properties();
-    p.setProperty("bootstrap.servers", kafkaContainer.getBootstrapServers());
-    p.setProperty("key.serializer", StringSerializer.class.getName());
-    p.setProperty("value.serializer", StringSerializer.class.getName());
-    KafkaProducer<String, String> producer = new KafkaProducer<>(p);
-    CompletableFuture.allOf(
-            sendMessage(producer, new ProducerRecord<>("oneTopic", 0, "theKey", "theValue")),
-            sendMessage(
-                producer, new ProducerRecord<>("twoTopic", 0, "anotherKey", "anotherValue")))
-        .get(10, TimeUnit.SECONDS);
-
-    producer.flush();
-
-    producer.close();
-    Dataset<Row> kafkaDf =
-        spark
-            .read()
-            .format("kafka")
-            .option("kafka.bootstrap.servers", kafkaContainer.getBootstrapServers())
-            .option("assign", "{\"oneTopic\": [0], \"twoTopic\": [0]}")
-            .load();
-    kafkaDf.collect();
-
-    StaticExecutionContextFactory.waitForExecutionEnd();
-    ArgumentCaptor<OpenLineage.RunEvent> lineageEvent =
-        ArgumentCaptor.forClass(OpenLineage.RunEvent.class);
-    Mockito.verify(SparkAgentTestExtension.OPEN_LINEAGE_SPARK_CONTEXT, times(4))
-        .emit(lineageEvent.capture());
-    OpenLineage.RunEvent completeEvent = lineageEvent.getAllValues().get(2);
-    assertThat(completeEvent).hasFieldOrPropertyWithValue(EVENT_TYPE, RunEvent.EventType.COMPLETE);
-    String kafkaNamespace =
-        "kafka://"
-            + kafkaContainer.getHost()
-            + ":"
-            + kafkaContainer.getMappedPort(KafkaContainer.KAFKA_PORT);
-    assertThat(completeEvent.getInputs())
-        .hasSize(2)
-        .satisfiesExactlyInAnyOrder(
-            dataset ->
-                assertThat(dataset)
-                    .hasFieldOrPropertyWithValue(NAME, "oneTopic")
-                    .hasFieldOrPropertyWithValue(NAMESPACE, kafkaNamespace),
-            dataset -> assertThat(dataset.getName()).isEqualTo("twoTopic"));
-  }
-
-  @Test
   @EnabledIfSystemProperty(named = "spark.version", matches = "(3.*)") // Spark version >= 3.*
   void testCacheReadFromFileWriteToParquet(@TempDir Path writeDir, SparkSession spark)
       throws InterruptedException, TimeoutException, IOException {
@@ -749,20 +631,5 @@ class SparkReadWriteIntegTest {
         .first()
         .hasFieldOrPropertyWithValue(NAMESPACE, bucketUrl)
         .hasFieldOrPropertyWithValue(NAME, "rdd_b");
-  }
-
-  private CompletableFuture sendMessage(
-      KafkaProducer<String, String> producer, ProducerRecord<String, String> record) {
-    CompletableFuture future = new CompletableFuture();
-    producer.send(
-        record,
-        (md, e) -> {
-          if (e != null) {
-            future.completeExceptionally(e);
-          } else {
-            future.complete(md);
-          }
-        });
-    return future;
   }
 }
